Move NumberInfo outside Charts
