+++ src/components/NumberInfo/index.js
