+++ src/components/NumberInfo/index.less
@@ -1,5 +1,5 @@
 @import "~antd/lib/style/themes/default.less";
-@import "../../../utils/utils.less";
+@import "../../utils/utils.less";
 
 .numberInfo {
   .suffix {
