+++ src/components/NumberInfo/index.md
@@ -1,0 +1,19 @@
+---
+type: General
+title: NumberInfo
+subtitle: 数据文本
+cols: 1
+---
+
+常用在数据卡片中，用于突出展示某个业务数据。
+
+## API
+
+参数 | 说明 | 类型 | 默认值
+----|------|-----|------
+title | 标题 | ReactNode\|string | -
+subTitle | 子标题 | ReactNode\|string | -
+total | 总量 | ReactNode\|string | -
+status | 增加状态 | 'up'\/'down' | -
+theme | 状态样式 | string | 'light'
+gap | 设置数字和描述直接的间距（像素） | number | 8
